+++ pyproject.toml
@@ -32,7 +32,7 @@ requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
-include = ["decision_tree_src*"]
+include = ["decision_tree*"]
 exclude = ["notebooks*", "tests*", "docs*"]
